refactor: Extract cube filters
